Quantity data type updated

+++ chart-service/src/main/java/com/taltech/ecommerce/chartservice/dto/OrderItemDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 public class OrderItemDto {
     private String inventoryCode;
-    private Integer quantity;
+    private BigDecimal quantity;
     private BigDecimal price;
     private LocalDateTime insertDate;
     private LocalDateTime updateDate;
